Add files via upload
Create a Servlet-based application to demonstrate the online examination system
for multiple choice questions. A session should be maintained for duration of the
examination, if the time gets over or student complets the exam within the specified
period, a Result Servlet should display number of correct and wrong answer. The
forms/output should appear as follows: